Separate helper methods (login and auth_params) from the tests
